Fixes patching of git.get_changed_files.
get_changed_files was getting set to a MagicMock instance, and this was
affecting downstream tests, for which it would remain monkey-patched as
a MagicMock instance. This would cause all tests run after tests in
TestGitFileManager to use a mock in place of slap.git.get_changed_files
instead of the actual function.